// frontend/src/company.d.ts
export interface CompanyBalanceSheet {
  accessNumber: string;
  symbol: string;
  year: number;
  quarter: number;
  form: string;
  startDate: string;
  endDate: string;
  report: {
    bs: {
      concept: string;
      label: string;
      unit: string;
      value: number;
    }[];
  };
}